Added Laberinto.c
Laberinto.c is a program that shows a maze in a console and ask you to
put the initial position of the person who is inside it and will solve
it.
As a known bug, there are some positions that cannot be solved, this
program can only pass two times at the only coordinates. It depends on
the maze structure.
You can easily modify the maze:
# : wall
| : end